Added ant task : test Fixed byte reading issues due to bad signed / unsigned integers handling Added FlowDirection Symbol CrescentCity2.nwc works !!!

+++ src/org/soldomi/support/nwc175/NwcEnum.java
@@ -1,5 +1,5 @@
 package org.soldomi.support.nwc175;
 
 public interface NwcEnum {
-  public byte getCode();
+  public int getCode();
 }
